# Add performance gh action