--- a/Algorithm.CSharp/Benchmarks/CoarseFineUniverseSelectionBenchmark.cs
+++ b/Algorithm.CSharp/Benchmarks/CoarseFineUniverseSelectionBenchmark.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
  *
@@ -34,7 +34,7 @@ namespace QuantConnect.Algorithm.CSharp.Benchmarks
             UniverseSettings.Resolution = Resolution.Minute;
 
             SetStartDate(2017, 11, 01);
-            SetEndDate(2018, 01, 01);
+            SetEndDate(2018, 3, 01);
             SetCash(50000);
 
             AddUniverse(CoarseSelectionFunction, FineSelectionFunction);