--- a/Algorithm.CSharp/Benchmarks/EmptySingleSecuritySecondEquityBenchmark.cs
+++ b/Algorithm.CSharp/Benchmarks/EmptySingleSecuritySecondEquityBenchmark.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
  *
@@ -27,7 +27,7 @@ namespace QuantConnect.Algorithm.CSharp.Benchmarks
         public override void Initialize()
         {
             SetStartDate(2008, 01, 01);
-            SetEndDate(2009, 01, 01);
+            SetEndDate(2008, 06, 01);
             SetBenchmark(dt => 1m);
             AddEquity("SPY", Resolution.Second);
         }