--- a/Algorithm.CSharp/Benchmarks/HistoryRequestBenchmark.cs
+++ b/Algorithm.CSharp/Benchmarks/HistoryRequestBenchmark.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
  *
@@ -23,7 +23,7 @@ namespace QuantConnect.Algorithm.CSharp.Benchmarks
         public override void Initialize()
         {
             SetStartDate(2010, 01, 01);
-            SetEndDate(2018, 01, 01);
+            SetEndDate(2022, 01, 01);
             SetCash(10000);
             _symbol = AddEquity("SPY").Symbol;
         }