--- a/Algorithm.CSharp/Benchmarks/IndicatorRibbonBenchmark.cs
+++ b/Algorithm.CSharp/Benchmarks/IndicatorRibbonBenchmark.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
  *
@@ -32,7 +32,7 @@ namespace QuantConnect.Algorithm.CSharp.Benchmarks
         public override void Initialize()
         {
             SetStartDate(2010, 01, 01);
-            SetEndDate(2018, 01, 01);
+            SetEndDate(2022, 01, 01);
 
             AddSecurity(SecurityType.Equity, "SPY", Resolution.Minute);