--- a/Algorithm.Python/Benchmarks/BasicTemplateBenchmark.py
+++ b/Algorithm.Python/Benchmarks/BasicTemplateBenchmark.py
@@ -23,7 +23,7 @@ class BasicTemplateBenchmark(QCAlgorithm):
 
     def Initialize(self):
         self.SetStartDate(2000, 1, 1)
-        self.SetEndDate(2017, 1, 1)    
+        self.SetEndDate(2022, 1, 1)
         self.SetBenchmark(lambda x: 1)
         self.AddEquity("SPY")