… Algorithm.Python/Benchmarks/EmptyEquityAndOptions500Benchmark.py