@Martin-Molinero 2022-07-20
```
MASTER
BasicTemplateBenchmark language CSharp avg dps: 254k samples: [254] avg length 10 sec
CoarseFineUniverseSelectionBenchmark language CSharp avg dps: 266k samples: [266] avg length 5 sec
EmptyEquityAndOptions500Benchmark language CSharp avg dps: 1012k samples: [1012] avg length 88 sec
EmptyMinute400EquityBenchmark language CSharp avg dps: 854k samples: [854] avg length 10 sec
EmptySingleSecuritySecondEquityBenchmark language CSharp avg dps: 248k samples: [248] avg length 47 sec
HistoryRequestBenchmark language CSharp avg dps: 178k samples: [178] avg length 10 sec
IndicatorRibbonBenchmark language CSharp avg dps: 171k samples: [171] avg length 9 sec
ScheduledEventsBenchmark language CSharp avg dps: 272k samples: [272] avg length 5 sec
StatefulCoarseUniverseSelectionBenchmark language CSharp avg dps: 147k samples: [147] avg length 2 sec
StatelessCoarseUniverseSelectionBenchmark language CSharp avg dps: 156k samples: [156] avg length 2 sec
BasicTemplateBenchmark language Python avg dps: 142k samples: [142] avg length 18 sec
CoarseFineUniverseSelectionBenchmark language Python avg dps: 259k samples: [259] avg length 6 sec
EmptyEquityAndOptions500Benchmark language Python avg dps: 963k samples: [963] avg length 92 sec
EmptyMinute400EquityBenchmark language Python avg dps: 772k samples: [772] avg length 11 sec
EmptySingleSecuritySecondEquityBenchmark language Python avg dps: 205k samples: [205] avg length 57 sec
HistoryRequestBenchmark language Python avg dps: 52k samples: [52] avg length 35 sec
IndicatorRibbonBenchmark language Python avg dps: 29k samples: [29] avg length 54 sec
ScheduledEventsBenchmark language Python avg dps: 161k samples: [161] avg length 8 sec
StatefulCoarseUniverseSelectionBenchmark language Python avg dps: 105k samples: [105] avg length 3 sec
StatelessCoarseUniverseSelectionBenchmark language Python avg dps: 112k samples: [112] avg length 2 sec

PR
BasicTemplateBenchmark language CSharp avg dps: 284k samples: [284] avg length 12 sec
CoarseFineUniverseSelectionBenchmark language CSharp avg dps: 340k samples: [340] avg length 9 sec
EmptyEquityAndOptions400Benchmark language CSharp avg dps: 1085k samples: [1085] avg length 68 sec
EmptyMinute400EquityBenchmark language CSharp avg dps: 903k samples: [903] avg length 19 sec
EmptySingleSecuritySecondEquityBenchmark language CSharp avg dps: 265k samples: [265] avg length 18 sec
HistoryRequestBenchmark language CSharp avg dps: 175k samples: [175] avg length 15 sec
IndicatorRibbonBenchmark language CSharp avg dps: 172k samples: [172] avg length 13 sec
ScheduledEventsBenchmark language CSharp avg dps: 282k samples: [282] avg length 7 sec
StatefulCoarseUniverseSelectionBenchmark language CSharp avg dps: 316k samples: [316] avg length 12 sec
StatelessCoarseUniverseSelectionBenchmark language CSharp avg dps: 334k samples: [334] avg length 11 sec
BasicTemplateBenchmark language Python avg dps: 149k samples: [149] avg length 24 sec
CoarseFineUniverseSelectionBenchmark language Python avg dps: 312k samples: [312] avg length 9 sec
EmptyEquityAndOptions400Benchmark language Python avg dps: 1022k samples: [1022] avg length 72 sec
EmptyMinute400EquityBenchmark language Python avg dps: 825k samples: [825] avg length 20 sec
EmptySingleSecuritySecondEquityBenchmark language Python avg dps: 206k samples: [206] avg length 23 sec
HistoryRequestBenchmark language Python avg dps: 51k samples: [51] avg length 35 sec
IndicatorRibbonBenchmark language Python avg dps: 30k samples: [30] avg length 52 sec
ScheduledEventsBenchmark language Python avg dps: 164k samples: [164] avg length 13 sec
StatefulCoarseUniverseSelectionBenchmark language Python avg dps: 160k samples: [160] avg length 24 sec
StatelessCoarseUniverseSelectionBenchmark language Python avg dps: 236k samples: [236] avg length 16 sec

```